Move sanitzeSchema method to schemaTree file
